refactor(hdfs-native): restructure HdfsNativeBackend and introduce HdfsNativeCore
